--- libs/ui/src/lib/team-card/team-card.component.html
@@ -1,6 +1,0 @@
-<section fxLayout="column" fxLayoutAlign="start center">
-  <img [imgRef]="userView.avatar" placeholderAsset="profil_user.webp" alt="User avatar">
-  <h5>{{ userView.name }}</h5>
-  <span class="mat-body-2">{{ userView.position }}</span>
-  <a class="mat-caption" [href]="'mailto:' + userView.email">{{ userView.email }}</a>
-</section>
