+++ libs/user/.storybook/config.js
@@ -1,0 +1,5 @@
+import { configure, addDecorator } from '@storybook/angular';
+import { withKnobs } from '@storybook/addon-knobs';
+
+addDecorator(withKnobs);
+configure(require.context('../src/lib', true, /\.stories\.tsx?$/), module);
