+++ libs/user/.storybook/webpack.config.js
@@ -1,0 +1,7 @@
+const rootWebpackConfig = require('../../../.storybook/webpack.config');
+// Export a function. Accept the base config as the only param.
+module.exports = async ({ config, mode }) => {
+  config = await rootWebpackConfig({ config, mode });
+  
+  return config;
+};
