+++ libs/user/src/lib/user/components/card/card.component.html
@@ -1,4 +1,6 @@
-<img [imgRef]="user.avatar" placeholderAsset="profil_user.webp" alt="User avatar" />
-<h5>{{ user.firstName }} {{ user.lastName }}</h5>
-<h4>{{ user.position }}</h4>
-<a class="mat-caption" [href]="'mailto:' + user.email">{{ user.email }}</a>
+<section fxLayout="column" fxLayoutAlign="start center">
+  <img [imgRef]="userView.avatar" placeholderAsset="profil_user.webp" alt="User avatar">
+  <h5>{{ userView.name }}</h5>
+  <span class="mat-body-2">{{ userView.position }}</span>
+  <a class="mat-caption" [href]="'mailto:' + userView.email">{{ userView.email }}</a>
+</section>
