+++ libs/user/src/lib/user/components/card/card.module.ts
@@ -1,9 +1,13 @@
+// Angular
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { CardComponent } from './card.component';
-import { ImageReferenceModule } from '@blockframes/ui/media/image-reference/image-reference.module';
 
+// Component
+import { CardComponent } from './card.component';
+
+// Blockframes
+import { ImageReferenceModule } from '@blockframes/ui/media/image-reference/image-reference.module';
 
 
 @NgModule({
