+++ nx.json
@@ -9,10 +9,16 @@
   },
   "projects": {
     "ui": {
-      "tags": ["ui", "shared"]
+      "tags": [
+        "ui",
+        "shared"
+      ]
     },
     "utils": {
-      "tags": ["utils", "shared"]
+      "tags": [
+        "utils",
+        "shared"
+      ]
     },
     "contract": {
       "tags": []
@@ -24,10 +30,15 @@
       "tags": []
     },
     "ethers": {
-      "tags": ["ethereum", "smart-contract"]
+      "tags": [
+        "ethereum",
+        "smart-contract"
+      ]
     },
     "movie": {
-      "tags": ["movie"]
+      "tags": [
+        "movie"
+      ]
     },
     "organization": {
       "tags": []
@@ -39,26 +50,38 @@
       "tags": []
     },
     "notification": {
-      "tags": ["notification"]
+      "tags": [
+        "notification"
+      ]
     },
     "catalog": {
-      "tags": ["catalog"]
+      "tags": [
+        "catalog"
+      ]
     },
     "catalog-e2e": {
       "tags": []
     },
     "festival": {
-      "tags": ["festival"]
+      "tags": [
+        "festival"
+      ]
     },
     "festival-e2e": {
       "tags": [],
-      "implicitDependencies": ["festival"]
+      "implicitDependencies": [
+        "festival"
+      ]
     },
     "user": {
-      "tags": ["user"]
+      "tags": [
+        "user"
+      ]
     },
     "event": {
-      "tags": ["event"]
+      "tags": [
+        "event"
+      ]
     },
     "e2e": {
       "tags": []
@@ -71,7 +94,13 @@
     "default": {
       "runner": "@nrwl/workspace/tasks-runners/default",
       "options": {
-        "cacheableOperations": ["build", "lint", "test", "e2e"]
+        "cacheableOperations": [
+          "build",
+          "lint",
+          "test",
+          "e2e",
+          "build-storybook"
+        ]
       }
     }
   }
